+++ ledger/accountdb.go
@@ -204,10 +204,10 @@ func writeCatchpointStateString(ctx context.Context, tx *sql.Tx, stateName strin
 	return false, err
 }
 
-func (cp *catchpointTracker) databaseSize(tx *sql.Tx) (size uint64, err error) {
+/*func (cp *catchpointTracker) databaseSize(tx *sql.Tx) (size uint64, err error) {
 	err = tx.QueryRow("SELECT page_count * page_size as size FROM pragma_page_count(), pragma_page_size()").Scan(&size)
 	return
-}
+}*/
 
 func storeCatchpoint(tx *sql.Tx, round basics.Round, fileName string, catchpoint string, fileSize int64) (err error) {
 	_, err = tx.Exec("DELETE FROM storedcatchpoints WHERE round=?", round)
