+++ ledger/archival_test.go
@@ -80,10 +80,6 @@ func (wl *wrappedLedger) trackerLog() logging.Logger {
 	return wl.l.trackerLog()
 }
 
-func (wl *wrappedLedger) isCatchpointRound(round basics.Round) bool {
-	return false
-}
-
 func getInitState() (genesisInitState InitState) {
 	blk := bookkeeping.Block{}
 	blk.CurrentProtocol = protocol.ConsensusCurrentVersion
